feat: redesign desktop search form
- Complete redesign of the desktop search form to match mobile styling
- Replaced avatar image with Lucide UserRound icon for passengers
- Added custom PassengerSelect dropdown with full Tailwind styling
- Improved layout: swap button, departure/arrival inputs, passager selector, search button
- Prepared the form for future DatePicker integration
- Responsive, clean, and consistent design with mobile version
